chore: centralize delivery class
