--- orders/serializers/delivery_time_slot.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from ..models import DeliveryTimeSlot
-
-class DeliveryTimeSlotSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = DeliveryTimeSlot
-        fields = '__all__'
